.06 s

+++ Model_Challenge/terminalVelocity.m
@@ -1,4 +1,4 @@
 function v = terminalVelocity(data)
     n = int32(length(data) / 4);
-    v = sum(data(end-n+1:end)) / n;
+    v = sum(data(end-n+1:end)) / double(n);
 end
